Move arg-block report generation to own module
